Move ScreenIcon back to jp migration package
