+++ WordPress/src/main/java/org/wordpress/android/ui/main/jetpack/migration/compose/components/ScreenIcon.kt
@@ -1,4 +1,4 @@
-package org.wordpress.android.ui.compose.components
+package org.wordpress.android.ui.main.jetpack.migration.compose.components
 
 import androidx.annotation.DrawableRes
 import androidx.compose.foundation.Image
