+++ WordPress/src/main/java/org/wordpress/android/ui/main/jetpack/migration/compose/state/DoneStep.kt
@@ -25,7 +25,6 @@ import androidx.compose.ui.unit.sp
 import org.wordpress.android.R
 import org.wordpress.android.ui.compose.components.ButtonsColumn
 import org.wordpress.android.ui.compose.components.PrimaryButton
-import org.wordpress.android.ui.compose.components.ScreenIcon
 import org.wordpress.android.ui.compose.components.Subtitle
 import org.wordpress.android.ui.compose.components.Title
 import org.wordpress.android.ui.compose.theme.AppTheme
@@ -33,6 +32,7 @@ import org.wordpress.android.ui.compose.utils.htmlToAnnotatedString
 import org.wordpress.android.ui.compose.utils.uiStringText
 import org.wordpress.android.ui.main.jetpack.migration.JetpackMigrationViewModel.ActionButton.DonePrimaryButton
 import org.wordpress.android.ui.main.jetpack.migration.JetpackMigrationViewModel.UiState
+import org.wordpress.android.ui.main.jetpack.migration.compose.components.ScreenIcon
 
 @Composable
 fun DoneStep(uiState: UiState.Content.Done) = with(uiState) {
