+++ WordPress/src/main/java/org/wordpress/android/ui/main/jetpack/migration/compose/state/ErrorStep.kt
@@ -12,7 +12,6 @@ import androidx.compose.ui.tooling.preview.Preview
 import org.wordpress.android.ui.compose.components.ButtonsColumn
 import org.wordpress.android.ui.compose.components.Message
 import org.wordpress.android.ui.compose.components.PrimaryButton
-import org.wordpress.android.ui.compose.components.ScreenIcon
 import org.wordpress.android.ui.compose.components.SecondaryButton
 import org.wordpress.android.ui.compose.components.Subtitle
 import org.wordpress.android.ui.compose.components.Title
@@ -21,6 +20,7 @@ import org.wordpress.android.ui.compose.utils.uiStringText
 import org.wordpress.android.ui.main.jetpack.migration.JetpackMigrationViewModel.ActionButton.ErrorPrimaryButton
 import org.wordpress.android.ui.main.jetpack.migration.JetpackMigrationViewModel.ActionButton.ErrorSecondaryButton
 import org.wordpress.android.ui.main.jetpack.migration.JetpackMigrationViewModel.UiState
+import org.wordpress.android.ui.main.jetpack.migration.compose.components.ScreenIcon
 import org.wordpress.android.ui.main.jetpack.migration.compose.dimmed
 
 @Composable
